uri: rework struct uri to use a copy of the original splitted string.
In the future, it is planned to extend the URI structure to allow its
passing with different options and in different formats. For these
purposes, it is planned to use functions that modify the source string,
for example `strtok_r`, so we need to rework the URI structure to create
copies of the string for each of the URI components.

Part of #5928