box: implement ability to pass several URIs in different ways
Previously, for "listen" and "replication" options URI can be passed
as a number, as a string value or as a table, which contains numbers
and strings, no special properties for URIs was available. Now user
can pass URIs in different ways, same as in tarantool "uri" library:
as before, as a table which contains URI and it's parameters in "param"
table, as a table which contains URI strings and URI tables. Same
as in "uri" library, which internally used to parse new "listen" and
"replication" there are different ways to specify properties for URI:
in a string which contains URI, after '?' delimiter, in a table which
contains URI in "params" table, in "default_params" table if it is
default parameters for all URIs.

Closes #5928