lua_cjson: add json-enc-slash option to compat
For unknown reason in upstream lua_cjson '/' was escaped
while according to the standard [rfc4627] it is unnecessary and is
questionably compatible with other implementations.

It was decided that the change will be introduced using
tarantool.compat (gh-7000). The patch adds json-escape-forward-slash
option to compat and its logic in lua_cjson and msgpuck.

Requires changes in tarantool-msgpuck [1] and tarantool.compat [2].
[1]: https://github.com/tarantool/msgpuck/pull/21
[2]: https://github.com/tarantool/tarantool/pull/7060

Fixes #6200
See also #7000
NO_DOC=docs on tarantool.wiki